pyproject.toml
[project]
name = "projeto-01-ingestao"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.84",
    "dlt[parquet,s3]>=1.24.0",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
